A-MoreOOP: Split functions into separate files
